remove references to style members from the parser classes
There is a change in 1.19.1 that breaks the parser (precision member
changed from plain number to an object). While modifying the parser to
be able to try different versions of the API is possible, it is quite a
task, and by simple omission of 'astyle' members the parser library
ignores the incoming data.

This (not providing 'astyle') also seems to work when adding new
transactions, at least with hledger-web 1.16.2